Fix compiler warning

+++ eternity/source/sdl/i_video.c
@@ -522,7 +522,7 @@ static void CenterMouse(void)
 {
    // Warp the the screen center
    
-   SDL_WarpMouse(video.width / 2, video.height / 2);
+   SDL_WarpMouse((short)(video.width / 2), (short)(video.height / 2));
    
    // Clear any relative movement caused by warping
